# BE-748 | Migrate from big.Float to float64 for balancer pools