This PR migrates from `big.Float` to `float64` for balancer pools. The  `float64` should have 52 bits of precision which should be more than enough and that in turn even further improves computations for balancer pools. Work is based on https://github.com/osmosis-labs/sqs/pull/657.

Previous:

```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase/pools/balancer
cpu: AMD Ryzen 7 5700X3D 8-Core Processor
BenchmarkSolveConstantFunctionInvariant
BenchmarkSolveConstantFunctionInvariant-16       2535862               473.4 ns/op
PASS
ok      github.com/osmosis-labs/sqs/router/usecase/pools/balancer       1.752s
```

New:

```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase/pools/balancer
cpu: AMD Ryzen 7 5700X3D 8-Core Processor
BenchmarkSolveConstantFunctionInvariant
BenchmarkSolveConstantFunctionInvariant-16      275088930                4.540 ns/op
PASS
ok      github.com/osmosis-labs/sqs/router/usecase/pools/balancer       1.762s
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Refactor**
  - Simplified arithmetic operations throughout the application by transitioning from high-precision number handling to standard floating-point calculations.
  - Improved performance and reduced complexity by replacing specialized numeric libraries with native floating-point types.
- **Bug Fixes**
  - Enhanced error handling for numeric parsing and calculations, reducing the risk of unexpected failures during operations involving pool parameters and token amounts.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->